Problem: "generic" nodes contained class data
All the nodes should only use the data members from the Node class, in
order to make the compiler flows more predictable.

GenericClass, GenericFunction and SprConcept classes contained an
InstantiationsSet contained some members, in additional to the
structure defined in the Node base-class. Stored this data as regular
children and referred nodes.

As Instantiation and InstantiationsSet were not nodes, transformed them
to subclass from Node too.