Problem: no possibility to visualise the AST of Sparrow code
Added a serialization functionality to the compiler. One can use save
node structures corresponding to source-codes before and after they are
compiled. The dumps are made both in binary and json format. The
reading functionality is not properly implemented.

Added a tool that transforms the dumped jsons into graphviz dot files
for visualization. The tool can then generate .pdf from the generated
.dot file and open the result.

In principle, the serialization functionality can be used to save the
entire program structure (e.g., for implementing separate compilation).
Still, there is still work to be done on that direction.