Problem: Indetifier and CompoundExp expand to useless Nop
Made these two nodes not to expand to a DeclExp by default. Only if the
form a callable expression, they will return a fun call.

Currently we set the DeclExp also as a property of the node itself. We
need to change this in the future. What prevents us for easily removing
this property is the handling of identifiers (or compound expressions
that are actually function calls.