Problem: identifiers would add a useless node property
Instead of adding a property with the referred declaration nodes, we
would expand to an DeclExp. In order to distinguish between the cases
in which we need to return an DeclExp or a function call we added a new
bool property to the Identifier and CompundExp nodes. Whenever we need
declarations, we can simply set that property to the base node

Also added support to pass properties to their explanation. For
example, we need to pass the newly added spr.allowDeclExp property from
a infix expression to a compound expression.