Problem: Type classes were still using virtual functions
Removed the virtual functions from the Type classes. For this we need
to  change the creation of the nodes. Now, each Type class is only a
wrapper around a stock TypeData object.

The changing of the mode is done now in a single function inside
Feather. As we also need this for ConceptType, this is not the ideal
place. To be fixed later