Problem: Creation of Feather general nodes still uses DynNodes
When creating Feather general nodes, we now use the standard mechanism of registering a node kind with some given functions.
This way, we avoid the usage of these DynNodes. This way we switch the compilation of these nodes from methods in the corresponding classes to our functions.

Theoretically we don't need the classes anymore, but we still have a lot of code that uses them. Mainly for retrieval purposes. To be removed later.