Problem: Ad-hoc initialisation of different systems
Created a module system. Each component will expose a module. then the module will be called to initialise itself, when the backend is initialised and when the compiler is destroyed. Each module can perform its own initialisation work.

Currently the SparrowCompiler will manually gather all the modules in the system