Problem: error reporting was using exceptions
Removed the usage of exceptions from error reporting code. If we cannot compile a node, we return NULL indicating that the node could not be compiled. The caller must ensure that it's working with a correctly compiled node.

Most probably we will have places in the code in which this is not correctly handled. We would address those issues while refactoring the nodes.