Problem: Nest Api is not separated from the utility and private code
Created a rough separation between nest files. We have now the "Api", "Utils" and "Internal" folders for Nest.
In "Api" we store code that needs to be exposed to the Module API.
In "Utils" we store non-API code, but that can be useful for other modules (that use the same language)
In "Internal" we store Nest-specific code