Problem: Cannot create ctor-like functions with the same name as the class
These are required if we want to extract the constructors outside the class.
The main problem was when returning a type with the same name as the function we are returning from. This would be considered a cycle in the previous implementation. Now, we consider only the types the are valid.
Also, now we can have both functions and types when we write a type name; changed the code a little bit to deal with that.