Problem: Work-stealing doesn't work; we cannot move tasks to other threads
Implemented a work-stealing algorithm. If a worker doesn't have work to do, it tries to steal work from other workers.
This way, it's possible to move tasks from one thread to another, and thus execute them in parallel.
It seems that this method is able to scale up nicely (at least on 4 cores on my machine).