Problem: No idling for threads that don't have work to do
Whenever a thread doesn't have work, it will enter into a wait state. We do this after several attempts to get some work.
Whenever new tasks are spawned or enqueued into the system, we wake up the waiting threads.
Whenever we decrease the ref-count to zero for a task, we try to wake up the thread that is waiting on that condition.