Added generator combinators
The elements() combinator takes a range of values and generates an element from that range each time.
The oneOf() combinator takes a range of generators, and randomly selects one to be used for generating all the values
The frequency combinator is similar to oneOf(), but for each generator it takes also a frequency for choosing the generator.