Operator lookup will search near the generic classes
Previously, because we create an artificial compilation context around the gneric instantiations, operator lookup will search near the class, meaning in this small context.
Now, changed the lookup code to search near the class, but in the first compilation context with a node associated with it. Therefore we will go to the parent package context.