Made it possible to issue errors without any colors
Addeded a flag to the compiler to instruct it not to use colors.
The Sublime build file will use this flag to disable the use of colors.
Also, added a newline before the errors, so that we don't write the errors just after the name of the file. This improves error reporting, as previously we would see the filename twice.