Cannot easily operate with datastructs: cannot make it an alias of an reference, no default ctor from its impl data type
Made the datatype with equal create an inner data variable named "_data" of the given type. This way, we allowed datatypes to be of any type.
Implicitely added the initCtor modifier to the SprClass of the datatypes with given type, so that we have an automatic ctor for the data.
No more usage of "base class" to implement datatypes; remove the base class part completely from SprClass