Renamed top-level "package" with "module"
Prepared the move towards modules. We no longer support the top-level package construct, and we replace it with the module construct. However, we currently ignore the last part of the module, and behave like before with packages.

We made this change to get rid of the bulk sparrow source code changes.