Enhanced AST dumping
Now AST dumping takes a string as argument. The source codes to be dumped
will be filtered by the given string. We will write the dumpled AST near
the files that generate them. We make sure that in a dumped JSON we
always dump the content that comes only from one source code.

Fixed some string encoding problems.

Added a new python tool that lists the types for the nodes at a given
location. This tool can be used to determine type errors.