Fix imports
Made the "import =" syntax deprecated. Whenever somebody says "import X"
we try to put the symbols in the current package, if there is no
qualifier given. This way, it seems more natural.

Fixed most of the tests.

The only 2 failing tests are the ones requiring Prolog imports.

Par tests compile but don't run properly (I think it was like that before)

The "meta" logic is not properly treated (precedenceUtils, debug). The
compiler expects that some of the intrinsics to be in current package,
and this creates confusion sometimes.