Imports are not copuled with module anymore
An import directive can now appear in the middle of the file, even inside
packages.

Simplified a lot the modules & import code.

Whenever the import name has its context set, we add the import to the
compiler. Then, the import is expanded to usings (and possibly package)
when the importName is semantically checked.