Made the logic tests not fail anymore
Made the user defined source code work again.
Each time we import something, no matter if we import a Sparrow file
or another file, we always create a using.
Changed the error of the using that points to nothing to a warning.
Ignored this warning for the using that imports a file.

Other minor fixes.