Fix problem in postfix ++ and -- for ranges.
If the RetType for a range was a reference, the code would return a
reference to an object allocated on the stack.

Solved the problem by using the RetType for the temporary variable. This
way, we return exactly the value (reference or not) that was in the range
before increment/decrement.