Added heuristic to not generate calls to empty destructors
After semantically checking a function applied a simple heuristic to
determine whether the function has some code that can have side effects.

If a function is marked as not having side effects, we can discard the
function call. If the arguments of the function have no side effects, we
can discard the arguments translation too.