First take on external ctors
We now support defining ctors outside the class. Whenever we search to
call a ctor, we also try to search the context near the class we want to
call the ctor for, not just the inner context of it.

It doesn't work in all cases yet (e.g., generics), but it's a start.