Removed dependency on Conan
- upgraded to LLVM 5.0.0 (from 3.5.2)
- upgraded to Boost 1.66 (from 1.60)
- upgraded to CMake 3.10.0 (3.0.0)

This is just making sure that the compiler compiles. By switching to
LLVM 5.0.0, the JIT engine is gone, and the new MCJIT is not working the
same way. Currently the compiler crashes the first time it tries to CT-
evaluate an expression.

Fixing the crash is left for a follow-up commit.