Different strategy for auto-generated functions.
Instead of checking for the existence of some functions, we always
generate the default functions (ctors, dtor, ctorFromCt, =,==). The
problem with checking for the existence of these functions is that it’s
sometimes very complex to check. This is especially true in the case of
generics and concepts.

The current method assigns a negative “overload priority” to the
generated functions. If we find a matching function, we discard the
functions with smaller overload priorities. This happens during the
overload selection.

Example:
  fun ctor(this: @GenType, arg: @AnyType) {}
  // automatically generated
  fun ctor(this: @GenType[Int], other: @GenType[Int]) {}

In this case, we would never chose the auto-generated constructor in the
overload mechanism.