Optimize custom convert
Each time there is a converison problem, we previously tried to find converion ctors. This can be expensive.

Instead, we now check only the dest classes that have the "convert" modifier. This modifier needs to be manually added to the class. This way, we speed up the custom conversion.

The performance results are not that great: from 1m54s to 1m52s for the whole test suite.