Add first version of formatTool
formatTool is able to read a source file and get its tokens and AST and
contain all the lexical & syntactical information of the input file. The
tokens also contain the whitespace and comments.

Reformat the Sparrow parsing code to be more reusable.

Add examples of how to implement interfaces by hand.